Working on the json parse implementation
