+++ minomaly/common/json/p_json_utils.h
@@ -1,4 +1,7 @@
 #pragma once
+#include "p_json_error.h"
+#include <sstream>
+#include <string>
 
 namespace Mino::Json::Private
 {
@@ -18,14 +21,51 @@ struct PropertyImpl
 };
 
 template <typename T>
-T parseValue(std::stringstream& stream)
+struct Type
 {
-    return T{}; // TODO
-}
+};
 
-template <>
-int parseValue<int>(std::stringstream& stream)
+template <typename FwIt>
+struct ParseImpl
 {
-    return 0; // TODO
-}
+    enum ParseState
+    {
+        Default,
+        Key,
+        Value
+    };
+
+    template <typename T>
+    T parse(Type<T>, FwIt begin, FwIt end)
+    {
+        auto result = T{};
+        // TODO: create property pointer / name / type lookup table
+        state = ParseState::Default;
+        for (; begin != end; ++begin)
+        {
+            switch (state)
+            {
+            case ParseState::Default:
+                break;
+            case ParseState::Key:
+                break;
+            case ParseState::Value:
+                break;
+            }
+        }
+        return result;
+    }
+
+    int parse(Type<int>, FwIt begin, FwIt end)
+    {
+        auto stream = std::stringstream{};
+        /*for (; begin != end and *begin != ','; ++begin)
+        {
+            stream
+        }*/
+        return 0; // TODO
+    }
+
+    ParseState state = ParseState::Default;
+};
 }
